+++ libs/core/src/lib/dynamic-form-element/dynamic-form-element-expression-data.ts
@@ -4,5 +4,7 @@ export interface DynamicFormElementExpressionData {
   readonly root?: DynamicFormFieldExpressionData;
   readonly parent?: DynamicFormElementExpressionData;
   readonly parentField?: DynamicFormFieldExpressionData;
+  readonly id: string;
+  readonly hidden: boolean;
   readonly [key: string]: any;
 }