+++ libs/core/src/lib/dynamic-form-field/dynamic-form-field-expression-data.ts
@@ -1,10 +1,11 @@
 import { DynamicFormElementExpressionData } from '../dynamic-form-element/dynamic-form-element-expression-data';
 
 export interface DynamicFormFieldExpressionData extends DynamicFormElementExpressionData {
-  readonly id: string;
   readonly key: string;
   readonly index: number;
   readonly depth: number;
+  readonly disabled: boolean;
+  readonly readonly: boolean;
   readonly model: any;
   readonly value: any;
   readonly valid: any;