+++ libs/material/src/lib/dynamic-form-input/dynamic-form-textbox/dynamic-form-textbox.module.ts
@@ -3,7 +3,7 @@ import { NgModule } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
-import { DynamicFormConfigModule, DynamicFormElementModule, DynamicFormInputType } from '@dynamic-forms/core';
+import { DynamicFormConfigModule, DynamicFormElementModule, DynamicFormInputType, DynamicFormTextboxModule } from '@dynamic-forms/core';
 import { matDynamicFormLibrary } from '../../dynamic-form-library/dynamic-form-library';
 import { MatDynamicFormTextboxComponent } from './dynamic-form-textbox.component';
 
@@ -20,6 +20,7 @@ export const matDynamicFormTextboxType: DynamicFormInputType = {
     MatFormFieldModule,
     MatInputModule,
     DynamicFormElementModule,
+    DynamicFormTextboxModule,
     DynamicFormConfigModule.withInput(matDynamicFormTextboxType),
   ],
   declarations: [